requirements.txt
torch >= 2.0.0
pandas
torchvision
pillow
numpy
scikit-learn
matplotlib
argparse
opencv-python
tqdm
yaml
pyyaml